[Feat] 스크롤이 있는지 판단하는 커스텀 훅 생성
